in alpha3/alpha2, power is calculated for every channel
Taking the PSD of the grand average results in a jagged power spectrum,
resulting in incredibly low values when calculating power. A better
approach is to calculate the power and power ratios for every single
channel, and then average them at the end to get a final value.

Note that this is not quite stable yet. Still need to bring method
options (i.e., default, grandAvg, etc) to updated version of
cl_alpha3alpha2.